included check player state

+++ src/index.js
@@ -28,16 +28,16 @@ var cors = require('cors')
 app.use(cors())
 app.listen(port, () => {
     console.log(`Rittersporn listening on port:  ${port}`)
-  })
+})
 
 //play
 app.get('/play', async function(req, res) {
   let guildID = req.query.guildID
   let channelID = req.query.channelID
   let url = req.query.url
-  
-  await playMusic(guildID, channelID, url, '')
-
+  if(player.state.status != Playing){
+    await playMusic(guildID, channelID, url, '')
+  }
   res.send(`playing`)
 })
 
@@ -48,8 +48,9 @@ app.get('/search', async function(req, res) {
   let searchTerm = req.query.searchTerm
 
   const searched = await searchMusic(searchTerm)
-  await playMusic(guildID,channelID,'', searched)
-
+  if(player.state.status != Playing){
+    await playMusic(guildID,channelID,'', searched)
+  }
   res.send(`search & play`)
 })
 
@@ -143,10 +144,12 @@ client.on('messageCreate', async (message) => {
         })
     }
     else if(message.content.slice(0, 6) === '!play ') {    
-        var url = message.content.slice(6)       
-        await playMusic(message.guild.id, message.member.voice.channel.id, url, '')
-        //delete message to keep channel clean 
-        setTimeout(() => message.delete(), 10000)
+        var url = message.content.slice(6)  
+        if(player.state.status != Playing){     
+            await playMusic(message.guild.id, message.member.voice.channel.id, url, '')
+            //delete message to keep channel clean 
+            setTimeout(() => message.delete(), 10000)
+        }
     }
     else if(message.content === '!stop') {
         player.stop()
